+++ src/app/client/src/app/modules/search/components/home-search/home-search.component.ts
@@ -529,7 +529,8 @@ export class HomeSearchComponent implements OnInit, OnDestroy, AfterViewInit {
     this.selectedFilters = filterData;
     const _cacheTimeout = _.get(this.allTabData, 'metaData.cacheTimeout') || 3600000;
     /* istanbul ignore next */
-    if (this.cacheService.exists('searchFiltersAll') && Object.keys(filterData).length > 0 && !_.get(filterData, 'key')) {
+    if (this.cacheService.exists('searchFiltersAll') && Object.keys(filterData).length > 0 && !_.get(filterData, 'key') 
+    && _.get(this.activatedRoute, 'snapshot.queryParams.ignoreSavedFilter') !== 'true' ) {
       const _searchFilters = this.cacheService.get('searchFiltersAll');
       let _cacheFilters = {
         primaryCategory: [..._.intersection(filterData['primaryCategory'], _searchFilters['primaryCategory']), ..._.difference(filterData['primaryCategory'], _searchFilters['primaryCategory'])],
@@ -544,10 +545,14 @@ export class HomeSearchComponent implements OnInit, OnDestroy, AfterViewInit {
         if (_cacheFilters[key] && _cacheFilters[key].length == 0) delete _cacheFilters[key];
       }
       this.cacheService.set('searchFiltersAll', this.selectedFilters, { expires: Date.now() + _cacheTimeout });
-    } else if (!this.cacheService.exists('searchFiltersAll') && Object.keys(filterData).length > 0 && !_.get(filterData, 'key')) {
+    } else if (!this.cacheService.exists('searchFiltersAll') && Object.keys(filterData).length > 0 && !_.get(filterData, 'key') 
+    && _.get(this.activatedRoute, 'snapshot.queryParams.ignoreSavedFilter') !== 'true') {
       this.cacheService.set('searchFiltersAll', filterData, { expires: Date.now() + _cacheTimeout });
     } else {
-      this.cacheService.remove('searchFiltersAll');
+      if(_.get(this.activatedRoute, 'snapshot.queryParams.ignoreSavedFilter') === 'true'){
+      } else{
+        this.cacheService.remove('searchFiltersAll');
+      }
     }
     const defaultFilters = _.reduce(filters, (collector: any, element) => {
       if (element.code === 'board') {
